requirements.txt
# Core
python-dotenv==1.0.1
certifi==2024.2.2
httpx==0.27.0

# Groq LLM
groq==0.5.0

# LangChain ecosystem
langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38
langchain-groq==1.1.4

# FastAPI server
fastapi==0.110.0
uvicorn==0.29.0

# Utilities
pydantic==1.10.13
typing-extensions==4.10.0